test: 댓글 좋아요 테스트 분리
