+++ src/test/java/com/salmalteam/salmal/application/comment/CommentServiceTest.java
@@ -16,10 +16,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.salmalteam.salmal.comment.application.CommentService;
 import com.salmalteam.salmal.comment.entity.Comment;
 import com.salmalteam.salmal.comment.entity.CommentRepository;
-import com.salmalteam.salmal.comment.entity.like.CommentLikeRepository;
 import com.salmalteam.salmal.comment.entity.report.CommentReportRepository;
 import com.salmalteam.salmal.comment.exception.CommentException;
-import com.salmalteam.salmal.comment.exception.like.CommentLikeException;
 import com.salmalteam.salmal.comment.exception.report.CommentReportException;
 import com.salmalteam.salmal.member.application.MemberService;
 import com.salmalteam.salmal.member.entity.Member;
@@ -37,8 +35,6 @@ class CommentServiceTest {
 	MemberService memberService;
 	@Mock
 	CommentRepository commentRepository;
-	@Mock
-	CommentLikeRepository commentLikeRepository;
 	@Mock
 	CommentReportRepository commentReportRepository;
 
@@ -74,84 +70,6 @@ class CommentServiceTest {
 			// when & then
 			assertThatThrownBy(() -> commentService.updateComment(memberId, commentId, voteCommentUpdateRequest))
 				.isInstanceOf(CommentException.class);
-
-		}
-	}
-
-	@Nested
-	class 댓글_좋아요_테스트 {
-
-		@Test
-		void 좋아요할_댓글이_존재하지_않으면_예외가_발생한다() {
-			// given
-			final Long memberId = 1L;
-			final Long commentId = 1L;
-
-			// when & then
-			assertThatThrownBy(() -> commentService.likeComment(memberId, commentId))
-				.isInstanceOf(CommentException.class);
-
-			verify(memberService).findMemberById(eq(memberId));
-		}
-
-		@Test
-		void 이미_좋아요를_한_경우_예외가_발생한다() {
-
-			// given
-			final Long memberId = 1L;
-			final Long commentId = 1L;
-
-			final Member member = Member.createActivatedMember("xxx", "닉네임", "kakao", true);
-			final Vote vote = Vote.of("imageUrl", member);
-			final Comment comment = Comment.of("댓글", vote, member);
-
-			given(memberService.findMemberById(any())).willReturn(member);
-			given(commentRepository.findById(any())).willReturn(Optional.of(comment));
-			given(commentLikeRepository.existsByCommentAndLiker(any(), any())).willReturn(true);
-
-			// when & then
-			assertThatThrownBy(() -> commentService.likeComment(memberId, commentId))
-				.isInstanceOf(CommentLikeException.class);
-
-		}
-
-	}
-
-	@Nested
-	class 댓글_좋아요_취소_테스트 {
-
-		@Test
-		void 좋아요_취소할_댓글이_존재하지_않으면_예외가_발생한다() {
-
-			// given
-			final Long memberId = 1L;
-			final Long commentId = 1L;
-
-			// when & then
-			assertThatThrownBy(() -> commentService.unLikeComment(memberId, commentId))
-				.isInstanceOf(CommentException.class);
-
-			verify(memberService).findMemberById(eq(memberId));
-		}
-
-		@Test
-		void 좋아요를_한_적이_없을_경우_예외가_발생한다() {
-
-			// given
-			final Long memberId = 1L;
-			final Long commentId = 1L;
-
-			final Member member = Member.createActivatedMember("xxx", "닉네임", "kakao", true);
-			final Vote vote = Vote.of("imageUrl", member);
-			final Comment comment = Comment.of("댓글", vote, member);
-
-			given(memberService.findMemberById(any())).willReturn(member);
-			given(commentRepository.findById(any())).willReturn(Optional.of(comment));
-			given(commentLikeRepository.existsByCommentAndLiker(any(), any())).willReturn(false);
-
-			// when & then
-			assertThatThrownBy(() -> commentService.unLikeComment(memberId, commentId))
-				.isInstanceOf(CommentLikeException.class);
 
 		}
 	}
